requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.8.0
asyncio-throttle==1.0.2
asyncpg==0.30.0
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.2
click==8.1.8
codebase-to-text==1.0.7
distro==1.9.0
fastapi==0.115.6
Flask==3.1.1
frozenlist==1.7.0
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.2.4
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
lxml==5.3.1
MarkupSafe==3.0.2
multidict==6.6.3
numpy==2.3.2
openai==1.68.2
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.1
playwright==1.55.0
pluggy==1.6.0
propcache==0.3.2
psycopg2==2.9.10
psycopg2-binary==2.9.10
pydantic==2.10.5
pydantic_core==2.27.2
pyee==13.0.0
Pygments==2.19.2
pyperclip==1.9.0
PySocks==1.7.1
pytest==8.4.1
pytest-asyncio==1.1.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
pytz==2025.2
ratelimit==2.2.1
requests==2.32.4
selenium==4.35.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
SQLAlchemy==2.0.41
starlette==0.41.3
structlog==25.4.0
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.34.0
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
yarl==1.20.1
